requirements.txt
PyYAML==5.1.1
gti-spe-lib-json-model